实现评论功能，未修复二级评论动态渲染Bug

+++ src/main/java/com/lyd/itshequ/controller/PublishController.java
@@ -59,7 +59,7 @@ public class PublishController {
 	}
 
 	@GetMapping("/publish/{id}")
-	public String edit(@PathVariable("id")Integer id,Model model){
+	public String edit(@PathVariable("id")Long id,Model model){
 		PostDTO postById = postService.getPostById(id);
 		model.addAttribute("post",postById);
 		return "publish";
