+++ src/main/java/com/lyd/itshequ/model/Comment.java
@@ -8,7 +8,7 @@ public class Comment {
     private Long parentId;
     private Integer type;
     private String commentValue;
-    private Integer commentator;
+    private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Long likeCount;
