+++ src/main/java/com/lyd/itshequ/service/PostService.java
@@ -5,11 +5,11 @@ import com.lyd.itshequ.bean.PostDTO;
 import com.lyd.itshequ.model.Post;
 
 public interface PostService {
-	PageDTO getPostByCreator(Integer id,Integer page, Integer pageSize);
+	PageDTO getPostByCreator(Long id,Integer page, Integer pageSize);
 	PageDTO getPostAll(Integer page, Integer pageSize);
-	PostDTO getPostById(Integer id);
+	PostDTO getPostById(Long id);
 
 	void createOrUpdate(Post post);
 
-    PostDTO incView(Integer id);
+    PostDTO incView(Long id);
 }
